URI escape custom field values when building a search uri.
This is an issue that is not isolated to the custom fields and we might
want this to be fixed for the other parts of the search query as well. As I
am not very familiar with the code and don't have a running RT setup I
am hesitant to escape the whole query. Please test fix and if anyone
with a working RT setup can test escaping the full URI: be my guest and
fix it :)

I added a test for this scenario as well, tests can be run with rake
spec.